move graph authentication provider closer to source
