+++ src/app/services/graph-client/index.ts
@@ -1,1 +1,4 @@
+import { GraphAuthenticationProvider } from './GraphAuthenticationProvider';
+
 export { GraphClient } from './graph-client';
+export const authProvider = new GraphAuthenticationProvider();
