+++ src/modules/authentication/index.ts
@@ -1,7 +1,4 @@
 import { AuthenticationWrapper } from './AuthenticationWrapper';
-import { GraphAuthenticationProvider } from "./GraphAuthenticationProvider";
 
-const authProvider = new GraphAuthenticationProvider();
-const authenticationWrapper = AuthenticationWrapper.getInstance();
+export const authenticationWrapper = AuthenticationWrapper.getInstance();
 
-export { authProvider, authenticationWrapper }
